refactor: Перенёс в отдельный пакет
