+++ src/main/java/ru/yandex/practicum/filmorate/dao/genre/GenreDao.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.dao.genre;
 
 import org.springframework.data.repository.CrudRepository;
 import ru.yandex.practicum.filmorate.model.Genre;
