+++ src/main/java/ru/yandex/practicum/filmorate/dao/review/ReviewDao.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.dao.review;
 
 import ru.yandex.practicum.filmorate.model.Review;
 import java.util.List;
