+++ src/main/java/ru/yandex/practicum/filmorate/dao/review/ReviewRatingDao.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.dao;
+package ru.yandex.practicum.filmorate.dao.review;
 
 import ru.yandex.practicum.filmorate.model.ReviewRating;
 import java.util.Optional;
