+++ src/main/java/ru/yandex/practicum/filmorate/dao/review/impl/ReviewRatingDaoImpl.java
@@ -1,10 +1,10 @@
-package ru.yandex.practicum.filmorate.dao.impl;
+package ru.yandex.practicum.filmorate.dao.review.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.dao.ReviewRatingDao;
+import ru.yandex.practicum.filmorate.dao.review.ReviewRatingDao;
 import ru.yandex.practicum.filmorate.mapper.ReviewRatingMapper;
 import ru.yandex.practicum.filmorate.model.ReviewRating;
 import java.util.Optional;
