+++ src/main/java/ru/yandex/practicum/filmorate/model/FilmSort.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate;
+package ru.yandex.practicum.filmorate.model;
 
 public enum FilmSort {
     FILMS_BY_DIRECTOR,
